removed unnecessary imports, removed corpus_name input, remove gutenberg references
